montant enchere double -> int

+++ src/main/java/fr/eni/projet/bo/Enchere.java
@@ -6,11 +6,11 @@ public class Enchere {
 	
 	int noEnchere;
 	LocalDate dateEnchere;
-	double montant_enchere;
+	int montant_enchere;
 	Utilisateur encherisseur;
 	ArticleVendu articleVendu;
 	
-	public Enchere(int noEnchere, LocalDate dateEnchere, double montant_enchere, Utilisateur encherisseur, ArticleVendu articleVendu) {
+	public Enchere(int noEnchere, LocalDate dateEnchere, int montant_enchere, Utilisateur encherisseur, ArticleVendu articleVendu) {
 		super();
 		this.noEnchere = noEnchere;
 		this.dateEnchere = dateEnchere;
@@ -19,7 +19,7 @@ public class Enchere {
 		this.articleVendu = articleVendu;
 	}
 
-	public Enchere(LocalDate dateEnchere, double montant_enchere, Utilisateur encherisseur, ArticleVendu articleVendu) {
+	public Enchere(LocalDate dateEnchere, int montant_enchere, Utilisateur encherisseur, ArticleVendu articleVendu) {
 		super();
 		this.dateEnchere = dateEnchere;
 		this.montant_enchere = montant_enchere;
@@ -43,11 +43,11 @@ public class Enchere {
 		this.dateEnchere = dateEnchere;
 	}
 
-	public double getMontant_enchere() {
+	public int getMontant_enchere() {
 		return montant_enchere;
 	}
 
-	public void setMontant_enchere(double montant_enchere) {
+	public void setMontant_enchere(int montant_enchere) {
 		this.montant_enchere = montant_enchere;
 	}
